requirements.txt
pyyaml
requests
dnspython
scikit-learn
numpy
tabulate